removing dependency on request-module introducing concept of response handlers to also support jsfiles and modules in the future
