+++ lib/components/asset-request-handler.js
@@ -31,7 +31,6 @@ function serveAssetFile(asset, res) {
 	console.log('Serving ' + asset);
 }
 
-
 var component = {
 
 	configure: function (hub, configuration) {
