+++ lib/components/proxy-request-handler.js
@@ -1,10 +1,13 @@
 'use strict';
 
 var httpProxy = require('http-proxy');
+var _ = require('lodash');
 var contracts = require('../../api/contracts');
 
 var targetAddress = 'http://localhost:8080/';
 var proxy;
+
+var responseHandlers = [];
 
 var component = {
 
@@ -19,6 +22,25 @@ var component = {
 		hub.provideService({
 			contract: contracts.requestHandler,
 			component: this
+		});
+		hub.requireService({
+			contract: contracts.responseHandler,
+			component: this,
+			optional: true,
+			aggregate: true,
+			bind: this.addResponseHandler,
+			unbind: this.removeResponseHandler
+		});
+	},
+	addResponseHandler: function (responseHandler) {
+		responseHandlers.push(responseHandler);
+		responseHandlers = _.sortBy(responseHandlers, function (responseHandler) {
+			return responseHandler.getPriority();
+		});
+	},
+	removeResponseHandler: function (responseHandler) {
+		_.remove(responseHandler, function (r) {
+			return r.getName() === responseHandler.getName();
 		});
 	},
 	start: function () {
@@ -44,6 +66,11 @@ var component = {
 		return true;
 	},
 	handle: function (req, res) {
+		_.forEach(responseHandlers, function(responseHandler) {
+			if(responseHandler.canProcess(req)) {
+				responseHandler.preProcess(res, req);
+			}
+		});
 		proxy.web(req, res, {target: targetAddress}, function (e) {
 			if (e.code === 'ECONNREFUSED') {
 				res.writeHead(504);
