+++ package.json
@@ -22,7 +22,6 @@
     "lodash": "^4.2.1",
     "http-proxy": "^1.12.0",
     "glob-all": "^3.0.1",
-    "request" : "^2.69.0",
     "properties-reader" : "^0.0.15"
   },
   "devDependencies": {
